refactor: add rect type and fix ~ imports
